requirements.txt
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.19
django-bootstrap4==3.0.1
django-chartjs==2.2.1
django-session-csrf==0.7.1
django-stdimage==5.3.0
django-weasyprint==1.1.0
gunicorn==20.0.4
html5lib==1.1
joblib==1.0.1
nltk==3.6.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
Pyphen==0.10.0
pytz==2021.1
regex==2021.4.4
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
textblob==0.15.3
tinycss2==1.1.0
tqdm==4.60.0
WeasyPrint==52.5
webencodings==0.5.1
